etl/ETL_Service/requirements.txt
annotated-types==0.7.0
backoff==2.2.1
black==22.1.0
certifi==2024.6.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.13.2
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.8.0
pydantic==2.7.3
pydantic-settings==2.3.0
pydantic_core==2.18.4
pyflakes==2.4.0
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.12.1
urllib3==2.2.1


